LayoutTests/proximity/check-event-deviceproximity-expected.txt

This test is for checking availablity of proximity event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'onwebkitdeviceproximity' in window is true
PASS document.createEvent('DeviceProximityEvent') instanceof window.DeviceProximityEvent is true
PASS document.createEvent('DeviceProximityEvent') instanceof window.Event is true
PASS document.createEvent('DeviceProximityEvent').constructor === window.DeviceProximityEvent is true
PASS successfullyParsed is true

TEST COMPLETE
